Changed to Minecraft 1.8.9 Updated mappings Implemented IInitListener (not sure yet, if it will stay like this)

+++ src/main/java/de/teamlapen/vampirism/config/BalanceVampirePlayer.java
@@ -38,8 +38,8 @@ public class BalanceVampirePlayer extends BalanceValues {
     public static double JUMP_TYPE;
     @DefaultInt(value = 3,name="Turn Others Level",minValue = 1,hasAlternate = true,alternateValue = 1,comment = "Level as of which a vampire player is able to infect other players")
     public static int MIN_TURN_LEVEL;
-    @DefaultDouble(value = 1.0F, name = "Blood Exhaustion Modifier", minValue = 0, maxValue = 5, comment = "Blood exhaustion is multiplied with this value")
-    public static int BLOOD_EXHAUSTION_MOD;
+    @DefaultDouble(value = 1.0D, name = "Blood Exhaustion Modifier", minValue = 0, maxValue = 5, comment = "Blood exhaustion is multiplied with this value")
+    public static double BLOOD_EXHAUSTION_MOD;
     @DefaultBoolean(value = true, name = "Blood Increase Exhaustion", comment = "Increase exhaustion modifier with higher levels")
     public static boolean BLOOD_INCREASE_EXHAUSTION;
